removed annotation based authrization

+++ src/main/java/dev/mm/core/coreservice/controller/UserController.java
@@ -1,11 +1,13 @@
 package dev.mm.core.coreservice.controller;
 
-import dev.mm.core.coreservice.annotations.RequiredRoles;
 import dev.mm.core.coreservice.dto.user.CreateUpdateUserDto;
 import dev.mm.core.coreservice.dto.user.UserPageRequestDto;
+import dev.mm.core.coreservice.security.AuthorizationService;
+import dev.mm.core.coreservice.security.UserDetailsImpl;
 import dev.mm.core.coreservice.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -14,7 +16,6 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import static dev.mm.core.coreservice.constants.Roles.ADMIN_ID;
 import static dev.mm.core.coreservice.constants.Uris.API_USER;
 import static dev.mm.core.coreservice.constants.Uris.API_USER_ID;
 
@@ -22,35 +23,48 @@ import static dev.mm.core.coreservice.constants.Uris.API_USER_ID;
 public class UserController {
 
     @Autowired
+    private AuthorizationService authorizationService;
+
+    @Autowired
     private UserService userService;
 
     @GetMapping(API_USER)
-    @RequiredRoles({ADMIN_ID})
-    public ResponseEntity getUsersPage(UserPageRequestDto userPageRequestDto) {
+    public ResponseEntity getUsersPage(
+        UserPageRequestDto userPageRequestDto,
+        @AuthenticationPrincipal UserDetailsImpl userDetails
+    ) {
+        authorizationService.ensureUserIsSuperAdmin(userDetails);
         return ResponseEntity.ok(userService.usersPage(userPageRequestDto));
     }
 
     @GetMapping(API_USER_ID)
-    @RequiredRoles({ADMIN_ID})
-    public ResponseEntity getUser(@PathVariable long userId) {
+    public ResponseEntity getUser(@PathVariable long userId, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        authorizationService.ensureUserIsSuperAdmin(userDetails);
         return ResponseEntity.ok(userService.getUserDtoById(userId));
     }
 
     @PostMapping(API_USER)
-    @RequiredRoles({ADMIN_ID})
-    public ResponseEntity createUser(@RequestBody CreateUpdateUserDto createUpdateUserDto) {
+    public ResponseEntity createUser(
+        @RequestBody CreateUpdateUserDto createUpdateUserDto,
+        @AuthenticationPrincipal UserDetailsImpl userDetails
+    ) {
+        authorizationService.ensureUserIsSuperAdmin(userDetails);
         return ResponseEntity.ok(userService.validateAndCreateUser(createUpdateUserDto));
     }
 
     @PutMapping(API_USER_ID)
-    @RequiredRoles({ADMIN_ID})
-    public ResponseEntity updateUser(@PathVariable long userId, @RequestBody CreateUpdateUserDto createUpdateUserDto) {
+    public ResponseEntity updateUser(
+        @PathVariable long userId,
+        @RequestBody CreateUpdateUserDto createUpdateUserDto,
+        @AuthenticationPrincipal UserDetailsImpl userDetails
+    ) {
+        authorizationService.ensureUserIsSuperAdmin(userDetails);
         return ResponseEntity.ok(userService.validateAndUpdateUser(userId, createUpdateUserDto));
     }
 
     @DeleteMapping(API_USER_ID)
-    @RequiredRoles({ADMIN_ID})
-    public ResponseEntity deleteUser(@PathVariable long userId) {
+    public ResponseEntity deleteUser(@PathVariable long userId, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        authorizationService.ensureUserIsSuperAdmin(userDetails);
         return ResponseEntity.ok(userService.deleteUser(userId));
     }
 }
